cmake 1.1 PG : attempt to support "muniversal" in the configure logs
Committed from host : Portia.local

+++ _resources/port1.0/group/cmake-1.1.tcl
@@ -502,7 +502,11 @@ proc cmake.save_configure_cmd {{save_log_too ""}} {
     if {${save_log_too} ne ""} {
         pre-configure {
             configure.pre_args-prepend "-cf '${configure.cmd} "
-            configure.post_args-append "|& tee ${workpath}/.macports.${subport}.configure.log'"
+            if {[info exists configure.build_arch]} {
+                configure.post_args-append "|& tee ${workpath}/.macports.${subport}.configure-${configure.build_arch}.log'"
+            } else {
+                configure.post_args-append "|& tee ${workpath}/.macports.${subport}.configure.log'"
+            }
             configure.cmd "/bin/csh"
             ui_debug "configure command set to `${configure.cmd} ${configure.pre_args} ${configure.args} ${configure.post_args}`"
         }
